fix numeric

+++ internal/migrator/migrations/000001_init.up.sql
@@ -12,7 +12,7 @@ CREATE TABLE
         user_id INT,
         number TEXT,
         process TEXT,
-        accrual INT,
+        accrual FLOAT (2),
         uploaded_at TEXT,
         FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
     );
@@ -21,6 +21,6 @@ CREATE TABLE
     IF NOT EXISTS balance (
         id SERIAL PRIMARY KEY,
         user_id INT,
-        balance FLOAT,
+        balance FLOAT (2),
         FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
     );
