+++ internal/migrator/migrations/000002_init.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE
         id SERIAL PRIMARY KEY,
         user_id INT,
         order_number TEXT,
-        sum INT,
+        sum FLOAT (2),
         processed_at TEXT,
         FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
     );
