+++ internal/store/update_balance.go
@@ -1,6 +1,6 @@
 package store
 
-func (store *Store) UpdateBalance(balance int, userID string) error {
+func (store *Store) UpdateBalance(balance float32, userID string) error {
 	_, err := store.db.Exec(updateBalanceQuery, balance, userID)
 
 	if err != nil {
